move ai grading into folder
